Extract attribute parsing code from diesel_derives
This commit puts the attribute parsing code from diesel_derives into a
seperate crate. The main motivation here is similar as with
diesel_table_macro_syntax, which is to be able to reuse the parsing
logic in other parts of the code base. For this the other main usage
might be in diesel-cli for the migration generation code.
