+++ Cargo.toml
@@ -33,7 +33,7 @@ members = [
     "examples/sqlite/relations",
     "examples/sqlite/wasm",
     "xtask",
-    "diesel_infer_query",
+    "diesel_infer_query", "diesel_attribute_parser",
 ]
 
 [workspace.package]
