+++ diesel_attribute_parser/src/deprecated/belongs_to.rs
@@ -3,8 +3,8 @@ use syn::token::Comma;
 use syn::{Ident, LitStr, parenthesized};
 
 use crate::deprecated::utils::parse_eq_and_lit_str;
+use crate::notes::BELONGS_TO_NOTE;
 use crate::parsers::BelongsTo;
-use crate::util::BELONGS_TO_NOTE;
 
 pub fn parse_belongs_to(name: Ident, input: ParseStream) -> Result<BelongsTo> {
     if input.is_empty() {
