+++ diesel_derives/src/lib.rs
@@ -27,11 +27,8 @@ use sql_function::ExternSqlBlock;
 use syn::parse_quote;
 
 mod allow_tables_to_appear_in_same_query;
-mod attrs;
-mod deprecated;
 mod field;
 mod model;
-mod parsers;
 mod util;
 
 mod as_changeset;
