+++ diesel_derives/src/queryable_by_name.rs
@@ -1,11 +1,12 @@
+use diesel_attribute_parser::AttributeSpanWrapper;
+use diesel_attribute_parser::CheckForBackend;
 use proc_macro2::{Span, TokenStream};
 use quote::quote;
 use syn::spanned::Spanned;
 use syn::{DeriveInput, Ident, LitStr, Result, Type, parse_quote, parse_quote_spanned};
 
-use crate::attrs::AttributeSpanWrapper;
 use crate::field::{Field, FieldName};
-use crate::model::{CheckForBackend, Model};
+use crate::model::Model;
 use crate::util::wrap_in_dummy_mod;
 
 pub fn derive(item: DeriveInput) -> Result<TokenStream> {
