+++ diesel_derives/src/util.rs
@@ -1,121 +1,9 @@
 use proc_macro2::TokenStream;
 use quote::quote;
-use syn::parse::{Parse, ParseStream, Peek, Result};
-use syn::token::Eq;
-use syn::{Data, DeriveInput, GenericArgument, Ident, Type, parenthesized, parse_quote};
+use syn::parse::Result;
+use syn::{Data, DeriveInput, GenericArgument, Type, parse_quote};
 
 use crate::model::Model;
-
-pub const COLUMN_NAME_NOTE: &str = "column_name = foo";
-pub const SQL_TYPE_NOTE: &str = "sql_type = Foo";
-pub const SERIALIZE_AS_NOTE: &str = "serialize_as = Foo";
-pub const DESERIALIZE_AS_NOTE: &str = "deserialize_as = Foo";
-pub const TABLE_NAME_NOTE: &str = "table_name = foo";
-pub const TREAT_NONE_AS_DEFAULT_VALUE_NOTE: &str = "treat_none_as_default_value = true";
-pub const TREAT_NONE_AS_NULL_NOTE: &str = "treat_none_as_null = true";
-pub const BELONGS_TO_NOTE: &str = "belongs_to(Foo, foreign_key = foo_id)";
-pub const MYSQL_TYPE_NOTE: &str = "mysql_type(name = \"foo\")";
-pub const SQLITE_TYPE_NOTE: &str = "sqlite_type(name = \"foo\")";
-pub const POSTGRES_TYPE_NOTE: &str = "postgres_type(name = \"foo\", schema = \"public\")";
-pub const POSTGRES_TYPE_NOTE_ID: &str = "postgres_type(oid = 37, array_oid = 54)";
-pub const SELECT_EXPRESSION_NOTE: &str =
-    "select_expression = schema::table_name::column_name.is_not_null()";
-pub const SELECT_EXPRESSION_TYPE_NOTE: &str =
-    "select_expression_type = dsl::IsNotNull<schema::table_name::column_name>";
-pub const CHECK_FOR_BACKEND_NOTE: &str = "diesel::pg::Pg";
-pub const BASE_QUERY_NOTE: &str =
-    "base_query = schema::table_name::table.order_by(schema::table_name::id)";
-pub const BASE_QUERY_TYPE_NOTE: &str =
-    "base_query_type = dsl::OrderBy<schema::table_name::table, schema::table_name::id>";
-pub const RENAME_ALL_NOTE: &str = "rename_all = \"camelCase\"";
-pub const RENAME_NOTE: &str = "rename = \"your_name\"";
-
-pub fn unknown_attribute(name: &Ident, valid: &[&str]) -> syn::Error {
-    let prefix = if valid.len() == 1 { "" } else { " one of" };
-
-    syn::Error::new(
-        name.span(),
-        format!(
-            "unknown attribute, expected{prefix} `{}`",
-            valid.join("`, `")
-        ),
-    )
-}
-
-pub fn parse_eq<T: Parse>(input: ParseStream, help: &str) -> Result<T> {
-    if input.is_empty() {
-        return Err(syn::Error::new(
-            input.span(),
-            format!(
-                "unexpected end of input, expected `=`\n\
-                 help: the correct format looks like `#[diesel({help})]`",
-            ),
-        ));
-    }
-
-    input.parse::<Eq>()?;
-    input.parse()
-}
-
-/// Specialized version of `parse_eq` for `syn::Type` with a customized error message for readability.
-/// This is useful because a great variety of tokens would be valid to parse as a `syn::Type`.
-pub fn parse_eq_type(input: ParseStream, help: &str) -> Result<Type> {
-    if input.is_empty() {
-        return Err(syn::Error::new(
-            input.span(),
-            format!(
-                "unexpected end of input, expected `=`\n\
-                 help: the correct format looks like `#[diesel({help})]`",
-            ),
-        ));
-    }
-
-    input.parse::<Eq>()?;
-    input
-        .parse::<Type>()
-        .map_err(|e| syn::Error::new(e.span(), "expected type"))
-}
-
-pub fn parse_paren<T: Parse>(input: ParseStream, help: &str) -> Result<T> {
-    if input.is_empty() {
-        return Err(syn::Error::new(
-            input.span(),
-            format!(
-                "unexpected end of input, expected parentheses\n\
-                 help: the correct format looks like `#[diesel({help})]`",
-            ),
-        ));
-    }
-
-    let content;
-    parenthesized!(content in input);
-    content.parse()
-}
-
-pub fn parse_paren_list<T, D>(
-    input: ParseStream,
-    help: &str,
-    sep: D,
-) -> Result<syn::punctuated::Punctuated<T, <D as Peek>::Token>>
-where
-    T: Parse,
-    D: Peek,
-    D::Token: Parse,
-{
-    if input.is_empty() {
-        return Err(syn::Error::new(
-            input.span(),
-            format!(
-                "unexpected end of input, expected parentheses\n\
-                 help: the correct format looks like `#[diesel({help})]`",
-            ),
-        ));
-    }
-
-    let content;
-    parenthesized!(content in input);
-    content.parse_terminated(T::parse, sep)
-}
 
 pub fn wrap_in_dummy_mod(item: TokenStream) -> TokenStream {
     quote! {
